+++ src/utils/connection/switch.tsx
@@ -3,7 +3,7 @@ import { Trans } from 'react-i18next';
 import { Observable, Observer } from 'rxjs';
 import Web3 from 'web3';
 import { EthereumChainConfig, MetamaskError, Network } from '../../model';
-import { findNetworkConfig, isNativeMetamaskChain } from './network';
+import { findNetworkConfig, isNativeMetamaskChain } from '../network/network';
 
 async function switchEthereumChain(network: Network): Promise<null> {
   const chain = findNetworkConfig(network) as EthereumChainConfig;
