+++ src/utils/index.ts
@@ -1,8 +1,8 @@
 export * from './airdrop';
 export * from './bridge';
+export * from './connection';
 export * from './helper';
 export * from './mmr';
 export * from './network';
-export * from './record';
-export * from './token';
+export * from './mappingToken';
 export * from './tx';
