+++ src/utils/network/index.ts
@@ -1,5 +1,5 @@
 export * from './network';
-export * from './switch';
-export * from './connection';
-export * from './entrance';
+export * from '../connection/switch';
+export * from '../connection/connection';
+export * from '../connection/entrance';
 export * from './graph';
