--- src/utils/token/index.ts
@@ -1,2 +1,0 @@
-export * from './tokenInfo';
-export * from './mappingToken';
