+++ src/utils/tx/index.ts
@@ -1,3 +1,2 @@
 export * from './common';
-export * from './issuing';
-export * from './redeem';
+export * from './sendExtrinsic';
